third_party/WebKit/LayoutTests/http/tests/devtools/console/console-trace-in-eval.html

<html>
<head>
<script src="../../inspector/inspector-test.js"></script>
<script src="../../inspector/console-test.js"></script>
<script>


function evalSource()
{
    function b()
    {
        console.trace();
    }

    function a()
    {
        b();
    }

    a();
}

function doEvalSource()
{
    setTimeout(function() {
        eval("(" + evalSource + ")()//# sourceURL=evalURL.js");
    }, 0);
}

function test() {
  function callback() {
    ConsoleTestRunner.dumpConsoleMessages();
    TestRunner.completeTest();
  }
  TestRunner.evaluateInPage('doEvalSource()');
  ConsoleTestRunner.addConsoleSniffer(callback);
}

</script>
</head>

<body onload="runTest()">
<p>
Tests that when console.trace is called in eval'ed script ending
with //# sourceURL=url it will dump a stack trace that will have
the url as the script source. <a href="https://bugs.webkit.org/show_bug.cgi?id=47252">Bug 47252.</a>
</p>

</body>
</html>
