third_party/WebKit/LayoutTests/http/tests/devtools/console/console-trace-in-eval-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that when console.trace is called in eval'ed script ending with //# sourceURL=url it will dump a stack trace that will have the url as the script source. Bug 47252.

evalURL.js:5 console.trace
b @ evalURL.js:5
a @ evalURL.js:10
evalSource @ evalURL.js:13
(anonymous) @ evalURL.js:14
(anonymous) @ console-trace-in-eval.js:29
setTimeout (async)
doEvalSource @ console-trace-in-eval.js:28
(anonymous) @ console-trace-in-eval.js:39
